PHPLIB-240: FindAndModify should throw for unsupported write concern
This reverts the logic in 21b7d92aef4def00abade10911bae41f1b5ea378, which was to ignore an unsupported write concern for the duration of 1.x. Since the Collection helper methods do not supply a default write concern option if it is not supported, this change will only affect users that were explicitly passing a write concern and relying on the library to ignore it.

While a minor BC break, this changes makes findAndModify consistent with other write commands.