Test that ChangeStream::next() resumes after dropped connection
The original intention of this test was to verify resumability in the event of a dropped connection; however, since socketTimeoutMS and maxAwaitTimeMS are fixed values, we cannot test a dropped connection followed by a successful resume. Instead, we will test that the driver attempts to resume once and only once before we expect a socket timeout after the second attempt.

Note: ChangeStream::rewind() does not currently resume (see: PHPLIB-322)