PHPLIB-712: Allow hint for unacknowledged writes using OP_MSG when supported by the server (#859)
* Sync unified CRUD spec tests

Synced with mongodb/specifications@59a5dadb2eeab8bba5a7d9a9acd47aa8c1d594cf

Note: aggregate tests are intentionally excluded since those will be synced by mongodb/mongo-php-library#854

* PHPLIB-711: Unwrap BulkWriteException when evaluating isClientError

* Client-side errors for hint with update, delete, and findAndModify

The CRUD spec requires a client-side error if hint is used and the server would not otherwise raise an error for an unsupported option. This was already being done, but this commit renames an internal constant and adds functional tests.

A client-side error is also required if hint is used with an unacknowledged write concern and the server does not support hint for the operation (regardless of error reporting for unsupported options). This was previously done only for FindAndModify; however, it failed to detect some acknowledged write concerns (PHPLIB-712). That issue has been fixed and the commit additionally adds this behavior for update and delete operations.