PHPLIB-592: Allow GridFS StreamWrapper to throw exceptions (#869)
Previously, exceptions were converted to E_USER_WARNING with a false return value for consistency with PHP's stream API (e.g. fread). This is not actually required and allowing the original exception to be propagated is likely more helpful for users and simplifies some handling in the spec test runner.