Remove `<?php` in code exmaples
Not sure if this works, but in the Symfony documentation we configured the syntax highlighting on a global level, so there is no need to add `<?php`, which also reduces some noise.
As you told me, there is some mongodb overall documentation stuff thingy, it might be a) not possible or b) it should be kept.

In this case, just close this PR, thanks.