Remove the words `easily` and `simply`
In Symfony we triy to be kind to newcomers, which means we try to avoid word like
"simply", "easily" etc., because it depends on the knowledge of the reader and it can therefore be offensive.

I am not sure how this project or especially MongoDB is handling such things, but from my POV it doesn't hurt. :shrug: