PHPLIB-1115: Fix waitForSnapshot() for MongoDB 7.0+ (#1083)
This fixes a bug in the original implementation where the method would return even if the query did not match a document. Instead, we should retry if either the query fails to match or a SnapshotUnavailable error is raised.

This issue was only discovered due to increased snapshot latency in MongoDB 7.0-dev. In previous versions, the snapshot was always available by the subsequent queries in testSnapshotQueries(), which uses waitForSnapshot().

Additionally, usleep() is added to avoid spamming the server with retry attempts. The TODO item for hrtime() was also removed, as there's no significant benefit to using it in this context and it's return types are not as convenient as microtime().