PHPLIB-476: Consider transaction readPreference in select_server (#1178)
This also refactors the conditionals in extract_session_from_options and extract_read_preference_from_options to improve readability.

select_server() previously did not consider the read preference of an active transaction. This isn't very significant, as transactions require a primary read preference, but it is correct to do so.

This also introduces select_server_for_write() to avoid inheriting an active transaction's readPreference option.