Retry once when disabling fail points (#1445)
The previous refactor in 90dcf186c8664a2cdd57314122d46b16234555c2 changed when we disable fail points. ManagesFailPointTrait now does so immediately after test operations are run instead of during tear down, so we may be hitting a point where the server stream was disconnected and libmongoc is not recovering (possibly related to CDRIVER-4532). Adding an extra retry attempt seems to overcome this.